changed the message for winner.
