+++ .cache/26cecec14f1e44d170666ceba20c6696.json
@@ -1,0 +1,1 @@
+{"id":"src/randomcolor.js","dependencies":[],"generated":{"js":"'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.getRandomColor = getRandomColor;\nfunction getRandomColor() {\n  var letters = '0123456789ABCDEF';\n  var color = '#';\n  for (var i = 0; i < 6; i++) {\n    color += letters[Math.floor(Math.random() * 16)];\n  }\n  return color;\n}","map":{"mappings":[{"name":"getRandomColor","generated":{"line":6,"column":8},"source":"src/randomcolor.js","original":{"line":1,"column":16}},{"name":"getRandomColor","generated":{"line":6,"column":25},"source":"src/randomcolor.js","original":{"line":1,"column":16}},{"generated":{"line":7,"column":0},"source":"src/randomcolor.js","original":{"line":1,"column":7}},{"name":"getRandomColor","generated":{"line":7,"column":9},"source":"src/randomcolor.js","original":{"line":1,"column":16}},{"generated":{"line":7,"column":23},"source":"src/randomcolor.js","original":{"line":1,"column":7}},{"generated":{"line":7,"column":26},"source":"src/randomcolor.js","original":{"line":1,"column":34}},{"generated":{"line":8,"column":0},"source":"src/randomcolor.js","original":{"line":2,"column":2}},{"name":"letters","generated":{"line":8,"column":6},"source":"src/randomcolor.js","original":{"line":2,"column":6}},{"generated":{"line":8,"column":16},"source":"src/randomcolor.js","original":{"line":2,"column":16}},{"generated":{"line":8,"column":34},"source":"src/randomcolor.js","original":{"line":2,"column":2}},{"generated":{"line":9,"column":0},"source":"src/randomcolor.js","original":{"line":3,"column":2}},{"name":"color","generated":{"line":9,"column":6},"source":"src/randomcolor.js","original":{"line":3,"column":6}},{"generated":{"line":9,"column":14},"source":"src/randomcolor.js","original":{"line":3,"column":14}},{"generated":{"line":9,"column":17},"source":"src/randomcolor.js","original":{"line":3,"column":2}},{"generated":{"line":10,"column":0},"source":"src/randomcolor.js","original":{"line":4,"column":2}},{"generated":{"line":10,"column":7},"source":"src/randomcolor.js","original":{"line":4,"column":7}},{"name":"i","generated":{"line":10,"column":11},"source":"src/randomcolor.js","original":{"line":4,"column":11}},{"generated":{"line":10,"column":15},"source":"src/randomcolor.js","original":{"line":4,"column":15}},{"generated":{"line":10,"column":16},"source":"src/randomcolor.js","original":{"line":4,"column":2}},{"name":"i","generated":{"line":10,"column":18},"source":"src/randomcolor.js","original":{"line":4,"column":18}},{"generated":{"line":10,"column":22},"source":"src/randomcolor.js","original":{"line":4,"column":22}},{"generated":{"line":10,"column":23},"source":"src/randomcolor.js","original":{"line":4,"column":2}},{"name":"i","generated":{"line":10,"column":25},"source":"src/randomcolor.js","original":{"line":4,"column":25}},{"generated":{"line":10,"column":28},"source":"src/randomcolor.js","original":{"line":4,"column":2}},{"generated":{"line":10,"column":30},"source":"src/randomcolor.js","original":{"line":4,"column":30}},{"name":"color","generated":{"line":11,"column":0},"source":"src/randomcolor.js","original":{"line":5,"column":4}},{"name":"letters","generated":{"line":11,"column":13},"source":"src/randomcolor.js","original":{"line":5,"column":13}},{"name":"Math","generated":{"line":11,"column":21},"source":"src/randomcolor.js","original":{"line":5,"column":21}},{"name":"floor","generated":{"line":11,"column":26},"source":"src/randomcolor.js","original":{"line":5,"column":26}},{"generated":{"line":11,"column":31},"source":"src/randomcolor.js","original":{"line":5,"column":21}},{"name":"Math","generated":{"line":11,"column":32},"source":"src/randomcolor.js","original":{"line":5,"column":32}},{"name":"random","generated":{"line":11,"column":37},"source":"src/randomcolor.js","original":{"line":5,"column":37}},{"generated":{"line":11,"column":43},"source":"src/randomcolor.js","original":{"line":5,"column":32}},{"generated":{"line":11,"column":48},"source":"src/randomcolor.js","original":{"line":5,"column":48}},{"generated":{"line":11,"column":50},"source":"src/randomcolor.js","original":{"line":5,"column":21}},{"generated":{"line":11,"column":51},"source":"src/randomcolor.js","original":{"line":5,"column":13}},{"generated":{"line":11,"column":52},"source":"src/randomcolor.js","original":{"line":5,"column":4}},{"generated":{"line":12,"column":0},"source":"src/randomcolor.js","original":{"line":6,"column":3}},{"generated":{"line":13,"column":0},"source":"src/randomcolor.js","original":{"line":7,"column":2}},{"name":"color","generated":{"line":13,"column":9},"source":"src/randomcolor.js","original":{"line":7,"column":9}},{"generated":{"line":13,"column":14},"source":"src/randomcolor.js","original":{"line":7,"column":2}},{"generated":{"line":14,"column":0},"source":"src/randomcolor.js","original":{"line":8,"column":1}}],"sources":{"src/randomcolor.js":"export function getRandomColor () {\n  var letters = '0123456789ABCDEF'\n  var color = '#'\n  for (var i = 0; i < 6; i++) {\n    color += letters[Math.floor(Math.random() * 16)]\n  }\n  return color\n}\n"},"lineCount":null}},"hash":"e14f84d20a6d8ca5bf463f9d624df76f","cacheData":{"env":{}}}
