+++ .cache/56e2a334fff2305e0a96e1717c4e3f48.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"./script.js","dynamic":true,"parent":"/Users/alexcorey/Desktop/Momentum/week_3/project-w3-fe-game-Alexswansboro/index.html","resolved":"/Users/alexcorey/Desktop/Momentum/week_3/project-w3-fe-game-Alexswansboro/script.js"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n  <meta charset=\"UTF-8\">\n  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n  <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n  <title>Document</title>\n</head>\n<body>\n  <style>\n    canvas{border:2px solid black;}\n  </style>\n  <canvas id=\"screen\" width=\"500\" height=\"500\">\n    <script src=\"/2651703dda5162006d65c04c6dfc050d.js\"></script>\n  </canvas>\n  \n</body>\n</html>\n"},"hash":"1b9ca9c7e94ed5e645ff192b3ddb8de8","cacheData":{}}
