+++ .cache/f606f5e43dab2940247c7dcbaaedd25b.json
@@ -1,0 +1,1 @@
+{"id":"src/constants.js","dependencies":[{"name":"./randomcolor","loc":{"line":2,"column":31},"parent":"/Users/alexcorey/Desktop/Momentum/week_3/project-w3-fe-game-Alexswansboro/src/constants.js","resolved":"/Users/alexcorey/Desktop/Momentum/week_3/project-w3-fe-game-Alexswansboro/src/randomcolor.js"}],"generated":{"js":"'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.colors = exports.screenSize = exports.context = exports.canvas = undefined;\n\nvar _randomcolor = require('./randomcolor');\n\nvar canvas = exports.canvas = document.getElementById('screen');\nvar context = exports.context = canvas.getContext('2d');\nvar screenSize = exports.screenSize = {\n  x: canvas.width,\n  y: canvas.height\n};\nvar colors = exports.colors = {\n  player: '#972D07',\n  bullets: '#ED1C24',\n  aliens: (0, _randomcolor.getRandomColor)()\n};","map":{"mappings":[{"generated":{"line":8,"column":0},"source":"src/constants.js","original":{"line":2,"column":0}},{"generated":{"line":10,"column":0},"source":"src/constants.js","original":{"line":4,"column":7}},{"name":"canvas","generated":{"line":10,"column":4},"source":"src/constants.js","original":{"line":4,"column":13}},{"name":"document","generated":{"line":10,"column":30},"source":"src/constants.js","original":{"line":4,"column":22}},{"name":"getElementById","generated":{"line":10,"column":39},"source":"src/constants.js","original":{"line":4,"column":31}},{"generated":{"line":10,"column":53},"source":"src/constants.js","original":{"line":4,"column":22}},{"generated":{"line":10,"column":54},"source":"src/constants.js","original":{"line":4,"column":46}},{"generated":{"line":10,"column":62},"source":"src/constants.js","original":{"line":4,"column":22}},{"generated":{"line":10,"column":63},"source":"src/constants.js","original":{"line":4,"column":7}},{"generated":{"line":11,"column":0},"source":"src/constants.js","original":{"line":5,"column":7}},{"name":"context","generated":{"line":11,"column":4},"source":"src/constants.js","original":{"line":5,"column":13}},{"name":"canvas","generated":{"line":11,"column":32},"source":"src/constants.js","original":{"line":5,"column":23}},{"name":"getContext","generated":{"line":11,"column":39},"source":"src/constants.js","original":{"line":5,"column":30}},{"generated":{"line":11,"column":49},"source":"src/constants.js","original":{"line":5,"column":23}},{"generated":{"line":11,"column":50},"source":"src/constants.js","original":{"line":5,"column":41}},{"generated":{"line":11,"column":54},"source":"src/constants.js","original":{"line":5,"column":23}},{"generated":{"line":11,"column":55},"source":"src/constants.js","original":{"line":5,"column":7}},{"generated":{"line":12,"column":0},"source":"src/constants.js","original":{"line":6,"column":7}},{"name":"screenSize","generated":{"line":12,"column":4},"source":"src/constants.js","original":{"line":6,"column":13}},{"generated":{"line":12,"column":38},"source":"src/constants.js","original":{"line":6,"column":26}},{"name":"x","generated":{"line":13,"column":0},"source":"src/constants.js","original":{"line":7,"column":2}},{"name":"canvas","generated":{"line":13,"column":5},"source":"src/constants.js","original":{"line":7,"column":5}},{"name":"width","generated":{"line":13,"column":12},"source":"src/constants.js","original":{"line":7,"column":12}},{"generated":{"line":13,"column":17},"source":"src/constants.js","original":{"line":6,"column":26}},{"name":"y","generated":{"line":14,"column":0},"source":"src/constants.js","original":{"line":8,"column":2}},{"name":"canvas","generated":{"line":14,"column":5},"source":"src/constants.js","original":{"line":8,"column":5}},{"name":"height","generated":{"line":14,"column":12},"source":"src/constants.js","original":{"line":8,"column":12}},{"generated":{"line":15,"column":0},"source":"src/constants.js","original":{"line":6,"column":26}},{"generated":{"line":15,"column":1},"source":"src/constants.js","original":{"line":6,"column":7}},{"generated":{"line":16,"column":0},"source":"src/constants.js","original":{"line":10,"column":7}},{"name":"colors","generated":{"line":16,"column":4},"source":"src/constants.js","original":{"line":10,"column":13}},{"generated":{"line":16,"column":30},"source":"src/constants.js","original":{"line":10,"column":22}},{"name":"player","generated":{"line":17,"column":0},"source":"src/constants.js","original":{"line":11,"column":2}},{"generated":{"line":17,"column":10},"source":"src/constants.js","original":{"line":11,"column":10}},{"generated":{"line":17,"column":19},"source":"src/constants.js","original":{"line":10,"column":22}},{"name":"bullets","generated":{"line":18,"column":0},"source":"src/constants.js","original":{"line":12,"column":2}},{"generated":{"line":18,"column":11},"source":"src/constants.js","original":{"line":12,"column":11}},{"generated":{"line":18,"column":20},"source":"src/constants.js","original":{"line":10,"column":22}},{"name":"aliens","generated":{"line":19,"column":0},"source":"src/constants.js","original":{"line":13,"column":2}},{"generated":{"line":19,"column":10},"source":"src/constants.js","original":{"line":13,"column":10}},{"generated":{"line":20,"column":0},"source":"src/constants.js","original":{"line":10,"column":22}},{"generated":{"line":20,"column":1},"source":"src/constants.js","original":{"line":10,"column":7}}],"sources":{"src/constants.js":"\nimport { getRandomColor } from './randomcolor'\n\nexport const canvas = document.getElementById('screen')\nexport const context = canvas.getContext('2d')\nexport const screenSize = {\n  x: canvas.width,\n  y: canvas.height\n}\nexport const colors = {\n  player: '#972D07',\n  bullets: '#ED1C24',\n  aliens: getRandomColor()\n}\n"},"lineCount":null}},"hash":"b1810959c6dbf74e05383441dbe2f4d8","cacheData":{"env":{}}}
