+++ src/Colliding.js
@@ -1,0 +1,10 @@
+
+export function colliding (b1, b2) {
+  return !(
+    b1 === b2 ||
+            b1.center.x + b1.size.x / 2 < b2.center.x - b2.size.x / 2 ||
+            b1.center.y + b1.size.y / 2 < b2.center.y - b2.size.y / 2 ||
+            b1.center.x - b1.size.x / 2 > b2.center.x + b2.size.x / 2 ||
+            b1.center.y - b1.size.y / 2 > b2.center.y + b2.size.y / 2
+  )
+}
